authentication added

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,20 @@
 import React, { useContext } from "react";
 import { WorkoutsContext } from "../contexts/WorkoutsContext";
+import { AuthContext } from "../contexts/ِِAuthContext";
 
 export default function WorkoutDetails({ workout }) {
   const { fetchWorkouts } = useContext(WorkoutsContext);
+  const { user } = useContext(AuthContext);
   const deleteWorkout = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
-      headers: { "Content-Type": "application/json" },
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     if (response.ok) {
       alert("deleted successfully");
